La gestion des erreurs a été enlevée de routes.py et mis sur le fichier à part errorhandler.py
